Move train_tokenizer to src/maxtext
